moved useUpdatedUserAssets hook from UI to the SDK
