+++ packages/ui/components/pages/Fuse/Modals/PoolModal/index.tsx
@@ -1,8 +1,8 @@
 import { Modal, ModalBody, ModalContent, ModalOverlay } from '@chakra-ui/react';
+import { FundOperationMode } from '@midas-capital/sdk';
 import { useEffect, useState } from 'react';
 
 import AmountSelect from '@ui/components/pages/Fuse/Modals/PoolModal/AmountSelect';
-import { FundOperationMode } from '@ui/constants/index';
 import { MarketData } from '@ui/hooks/useFusePoolData';
 
 interface PoolModalProps {
