+++ packages/ui/hooks/fuse/useUpdatedUserAssets.ts
@@ -1,10 +1,10 @@
-import { NativePricedFuseAsset } from '@midas-capital/sdk';
-import { BigNumber, constants, utils } from 'ethers';
+import { FundOperationMode } from '@midas-capital/sdk';
+import { BigNumber } from 'ethers';
 import { useMemo } from 'react';
 import { useQuery, UseQueryResult } from 'react-query';
 
-import { FundOperationMode } from '@ui/constants/index';
 import { useRari } from '@ui/context/RariContext';
+import { MarketData } from '@ui/hooks/useFusePoolData';
 import { useUSDPrice } from '@ui/hooks/useUSDPrice';
 
 // TODO Write proper tests and fix `Native` naming issue for values in Fiat USD.
@@ -14,7 +14,7 @@ interface UseUpdatedUserAssetsResult<T> {
   index: number;
   amount: BigNumber;
 }
-const useUpdatedUserAssets = <T extends NativePricedFuseAsset>({
+const useUpdatedUserAssets = <T extends MarketData>({
   mode,
   index,
   assets,
@@ -23,7 +23,7 @@ const useUpdatedUserAssets = <T extends NativePricedFuseAsset>({
   const { fuse, currentChain, coingeckoId } = useRari();
   const { data: usdPrice } = useUSDPrice(coingeckoId);
 
-  const { data: updatedAssets }: UseQueryResult<NativePricedFuseAsset[]> = useQuery(
+  const { data: updatedAssets }: UseQueryResult<MarketData[]> = useQuery(
     [
       'useUpdatedUserAssets',
       currentChain.id,
@@ -36,107 +36,22 @@ const useUpdatedUserAssets = <T extends NativePricedFuseAsset>({
     async () => {
       if (!assets || !assets.length || !usdPrice) return [];
 
-      const assetToBeUpdated = assets[index];
-
-      const interestRateModel = await fuse.getInterestRateModel(assetToBeUpdated.cToken);
-
-      let updatedAsset: NativePricedFuseAsset;
-
-      if (mode === FundOperationMode.SUPPLY) {
-        const supplyBalance = assetToBeUpdated.supplyBalance.add(amount);
-
-        const totalSupply = assetToBeUpdated.totalSupply.add(amount);
-
-        updatedAsset = {
-          ...assetToBeUpdated,
-
-          supplyBalance,
-          totalSupply,
-          supplyBalanceNative:
-            Number(utils.formatUnits(supplyBalance, 18)) *
-            Number(utils.formatUnits(assetToBeUpdated.underlyingPrice, 18)),
-
-          supplyRatePerBlock: interestRateModel.getSupplyRate(
-            totalSupply.gt(constants.Zero)
-              ? assetToBeUpdated.totalBorrow.mul(constants.WeiPerEther).div(totalSupply)
-              : constants.Zero
-          ),
-        };
-      } else if (mode === FundOperationMode.WITHDRAW) {
-        const supplyBalance = assetToBeUpdated.supplyBalance.sub(amount);
-
-        const totalSupply = assetToBeUpdated.totalSupply.sub(amount);
-
-        updatedAsset = {
-          ...assetToBeUpdated,
-
-          supplyBalance,
-          totalSupply,
-          supplyBalanceNative:
-            Number(utils.formatUnits(supplyBalance, 18)) *
-            Number(utils.formatUnits(assetToBeUpdated.underlyingPrice, 18)),
-
-          supplyRatePerBlock: interestRateModel.getSupplyRate(
-            totalSupply.gt(constants.Zero)
-              ? assetToBeUpdated.totalBorrow.mul(constants.WeiPerEther).div(totalSupply)
-              : constants.Zero
-          ),
-        };
-      } else if (mode === FundOperationMode.BORROW) {
-        const borrowBalance = assetToBeUpdated.borrowBalance.add(amount);
-
-        const totalBorrow = assetToBeUpdated.totalBorrow.add(amount);
-
-        updatedAsset = {
-          ...assetToBeUpdated,
-
-          borrowBalance,
-          totalBorrow,
-
-          borrowBalanceNative:
-            Number(utils.formatUnits(borrowBalance, 18)) *
-            Number(utils.formatUnits(assetToBeUpdated.underlyingPrice, 18)) *
-            usdPrice,
-
-          borrowRatePerBlock: interestRateModel.getBorrowRate(
-            assetToBeUpdated.totalSupply.gt(constants.Zero)
-              ? totalBorrow.mul(constants.WeiPerEther).div(assetToBeUpdated.totalSupply)
-              : constants.Zero
-          ),
-        };
-      } else if (mode === FundOperationMode.REPAY) {
-        const borrowBalance = assetToBeUpdated.borrowBalance.sub(amount);
-
-        const totalBorrow = assetToBeUpdated.totalBorrow.sub(amount);
-
-        const borrowRatePerBlock = interestRateModel.getBorrowRate(
-          assetToBeUpdated.totalSupply.gt(constants.Zero)
-            ? totalBorrow.mul(constants.WeiPerEther).div(assetToBeUpdated.totalSupply)
-            : constants.Zero
-        );
-
-        updatedAsset = {
-          ...assetToBeUpdated,
-
-          borrowBalance,
-          totalBorrow,
-
-          borrowBalanceNative:
-            Number(utils.formatUnits(borrowBalance)) *
-            Number(utils.formatUnits(assetToBeUpdated.underlyingPrice)) *
-            usdPrice,
-
-          borrowRatePerBlock,
-        };
+      const resAssets = await fuse.getUpdatedAssets(mode, index, assets, amount);
+      const assetsWithPrice: MarketData[] = [];
+      if (resAssets && resAssets.length !== 0) {
+        resAssets.map((asset) => {
+          assetsWithPrice.push({
+            ...asset,
+            supplyBalanceFiat: asset.supplyBalanceNative * usdPrice,
+            borrowBalanceFiat: asset.borrowBalanceNative * usdPrice,
+            totalSupplyFiat: asset.totalSupplyNative * usdPrice,
+            totalBorrowFiat: asset.totalBorrowNative * usdPrice,
+            liquidityFiat: asset.liquidityNative * usdPrice,
+          });
+        });
       }
 
-      return assets.map((value, _index) => {
-        if (_index === index) {
-          return updatedAsset;
-        } else {
-          return value;
-        }
-      });
+      return assetsWithPrice;
     }
   );
